CODEBASE/Vocal Feature Analysis/requirements.txt
# Core deep learning and audio processing
torch>=2.0.0
torchaudio>=2.0.0

# Diarization, Vocal Feature Extraction, and Transcription
pyannote.audio>=3.1.0
openai-whisper>=20231117
transformers>=4.30.0

# Hugging Face model hub access and data handling
huggingface_hub>=0.17.0
pandas>=2.0.0
numpy>=1.24.0

# Audio file handling and preprocessing
soundfile>=0.12.1
ffmpeg-python>=0.2.0
noisereduce>=3.0.0
scipy>=1.10.0

# Language detection
langdetect>=1.0.9
